+++ src/main/java/com/hashfunctions/Djb2.java
@@ -1,8 +1,8 @@
 package com.hashfunctions;
 
 public final class Djb2 {
-    public static long hash(String s) {
-        long hash = 5381;
+    public static int hash(String s) {
+        int hash = 5381;
         for(int i = 0; i < s.length(); ++i) {
             hash = ((hash << 5) + hash) + s.charAt(i);
         }
