+++ src/main/java/com/hashfunctions/LoseLose.java
@@ -1,8 +1,8 @@
 package com.hashfunctions;
 
 public final class LoseLose {
-    public static long hash(String s) {
-        long hash = 0;
+    public static int hash(String s) {
+        int hash = 0;
         for(int i = 0; i < s.length(); ++i) {
             hash += s.charAt(i);
         }
